Create isPerfect.js
The isPerfect() function, which takes a number and returns true if it is perfect, and false otherwise.

A perfect number is a positive integer equal to the sum of its positive divisors (not counting the number itself). The divisor is the number by which the original number is divided. For example, the number 6 has three divisors: 1, 2 and 3. The number 6 is divisible by any of these numbers. Also, 6 is an ideal number, because 6 = 1 + 2 + 3 .

isPerfect(6); // true
isPerfect(7); // false